Configuration Fixes
Fixed the bug found in the last commit, it was worst than just not
showing the right picture, the program never updated the configuration
file. The escribirUser() methos had a condition that prevented it from
updating; although removing the condition might  solve the problem,
RandomAccesFile was used rather than ObjectOuputStream to write the
file.