Super Commit
List of changes:
-Now one can add movies
-One edit their profile, all validation is made but it doesn't let the
user know what is going on. That should be fixed.
-Added methods for saving movies to file, all related enums were
modified in order for this to get done.
-Fiexd an incredibly SEVERE, CATASTROPHIC, DISASTROUS, SERIOUS, FATAL
and similar nouns that denote something very bad, in the UserBuilder's
leerUser() method. This bastard continued to read if the credential was
false when the username matched, wich in turn caused an EOFException.
Such bug took me more than half a fucking hour to find.
-A similar bug was found in the desactivarUser() method, it waas fixed
too.
-Some prints to the command line were suppressed.
-Some images were added to the resources.
-CSS files were moved, and changed any reference to them to match the
new location.
-A class for printing the movies was created.