New feature Addes: Salas
Now you can add Salas and they'll get written to disk. In the future
they will be read for the tickets. Lots of classes were modified in
order to make evrything work.
-SalaLayout now has a boolean indicating if the Sala is 3D or not. Its
constructor was modified to include this value.
-Some getters and setters were added.
-A method for retrieving the state of a Seat was created. This tells
wheter the seat is hidden, available, selected or reserved.
-A method for countin how many seats with a given state are in the Sala,
this'll be used when the number of available seats is needed in the
Clien module.
-Fiexd a bug in the toString method of SalaLayout, that didn't append
the letter of a seat but it overrode it.
-A toInt(0 method was added to the SeatState enum, it ill be used to
know the state of a seat in less lines of code.
-Code was written in the AdminWindowController to handle the Disenio
request.
-ChairGrid has been finished, it now works as expected. It is anchored
to the window, clicking a seat modifies its corresponding seat in the
SalaLayout.
-The Disenio helper class now has the necessary features for handling
the design of the SalaLayout, the constructor was fixed, new methods
were created, it now writes the SalaLayout to disk on click of the save
button.
-SillaControl: its type was changed, it is an AnchorPane now, it
contains a Label with the name of the chair. Deleted some not working
shit like the ObservableValue attribute and its 'implmentns
ChangeListener'. Now it changes its look based on the new state,
triggered by the setState() mehtod.
-A ListRooms class was created for printing the Layout of the Salas.